pyproject.toml
[project]
name = "mystudy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
]
